Intermediate commit
this seems to work both with the old examples and with phase-filed
materials, but I only tried the axial-symmetric example

I am makng this commit to have a return point while I am tryng to make
some major changes to the architecture. Namely I want to have a function
that makes the residual and the tangent stiffness matrix, when is given
the elements, the displacement vectors, and whatevere is needed, and I
want to orgnize all of the get\phi functions
